Revert "KYC Form Page"
