+++ package.json
@@ -10,7 +10,6 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "api": "^6.1.1",
     "axios": "^1.4.0",
     "bcryptjs-react": "^2.4.6",
     "blurhash": "^2.0.5",
