--- src/components/Dashboard/SalesAdmin.jsx
@@ -1,12 +1,0 @@
-import React, { useState, useEffect } from "react";
-
-// components
-import OverallSales from "./SalesAdmin/OverallSales";
-
-export default function Sales() {
-  return (
-    <>
-      <OverallSales />
-    </>
-  );
-}
