--- src/components/Dashboard/SalesAdmin/Css/OrderLayUpdate.module.css
@@ -1,122 +1,0 @@
-.mDiv {
-  position: fixed;
-  top: 0px;
-  left: 0px;
-  width: 100%;
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  backdrop-filter: blur(10px);
-  background: rgba(0, 0, 0, 0.2);
-  z-index: 3;
-}
-
-.contendDiv,
-.contendDivDel {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  background-color: #fff;
-  backdrop-filter: blur(0px);
-  padding: 30px 50px;
-  border-radius: 15px;
-  color: #000;
-  overflow-y: scroll;
-}
-
-.contendDiv {
-  margin-top: 44px;
-  width: 80%;
-  max-height: 80vh;
-}
-
-.tableCatTTagDel {
-  width: 100%;
-  border-collapse: collapse;
-  text-align: center;
-}
-
-.UpdateOrderPTag {
-  font-size: 20px;
-  font-weight: 600;
-}
-
-.closeBtn {
-  position: absolute;
-  top: 30px;
-  right: 50px;
-  cursor: pointer;
-  transition: all 0.3s;
-}
-
-.closeBtn:hover {
-  color: #880000;
-}
-
-.BtnDivMain {
-  margin: 30px 0px;
-  width: 100%;
-  display: flex;
-  justify-content: space-evenly;
-  align-items: center;
-}
-
-.BtnDiv {
-  padding: 7px 30px;
-  border-radius: 8px;
-  font-size: 18px;
-  font-weight: 500;
-  color: #fff;
-}
-
-#Accept {
-  background-color: #1f9254;
-}
-
-#InProgress {
-  background-color: #3f81e0;
-}
-
-#Completed {
-  background-color: #4bb543;
-}
-
-#Cancelled {
-  background-color: #ff0000;
-}
-
-.ProductDelTableDiv {
-  width: 100%;
-}
-
-.titleItemDiv {
-  display: flex;
-  justify-content: center;
-  align-content: center;
-}
-
-.itemImg {
-  width: 30px;
-  height: 30px;
-  border-radius: 50%;
-  object-fit: contain;
-  aspect-ratio: 1/1;
-  margin-right: 10px;
-}
-
-.ProductDelPTag {
-  font-size: 18px;
-  font-weight: 400;
-  margin: 0px 0px 20px 0px;
-}
-
-.tableCatTTagDel td {
-  padding: 27px 5px 27px;
-  border-bottom: 1px solid #ddd;
-}
-
-.tableCatTTagDel tr:hover {
-  background-color: #ebefff;
-}
