--- src/components/Dashboard/SalesAdmin/Css/overallSales.module.css
@@ -1,335 +1,0 @@
-.mDIvTabs {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 0px 10px;
-  overflow-x: scroll;
-  overflow-y: hidden;
-  padding: 0px 3px;
-  transition: all 0.3;
-  margin-bottom: 30px;
-}
-
-.thTag {
-  font-weight: bold;
-}
-
-.top {
-  font-weight: 500;
-}
-
-.mDIvTabs::-webkit-scrollbar {
-  display: none;
-}
-
-.mDIvSideTabs {
-  display: flex;
-}
-
-.stateTrTag {
-  text-align: center;
-}
-
-.filters {
-  display: flex;
-  margin-top: 50px;
-}
-
-.select {
-  width: 45%;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  gap: 0px 20px;
-}
-
-.selectInner {
-  width: 47%;
-  padding: 9px 20px;
-  border: 1px solid #ddd;
-  border-radius: 5px;
-}
-
-.selectInner select {
-  width: 100%;
-  font-size: 14px;
-  border: none;
-  outline: none;
-}
-
-.search {
-  width: 50%;
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-}
-
-.resetFilBtn {
-  margin-left: 20px;
-  cursor: pointer;
-}
-
-.search input {
-  width: 70%;
-  padding: 0px 20px;
-  border: none;
-  border-radius: 5px;
-  border: 1px solid#DDD;
-  outline: none;
-  padding: 12px 7px;
-}
-
-.middle {
-  border-top: 1px solid #f6f6f6;
-  display: inline-flex;
-  padding: 15px;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-  width: 100%;
-}
-
-.tabs div {
-  padding-bottom: 10px;
-  font-size: 16px;
-  cursor: pointer;
-}
-
-.table {
-  padding: 40px 30px;
-  overflow-x: scroll;
-  width: 100%;
-}
-
-.svgTag {
-  margin-left: 5px;
-}
-
-::-webkit-scrollbar {
-  display: none;
-}
-
-.table th {
-  padding-block: 5px;
-  text-align: left;
-  border: 2px solid white;
-}
-
-.table table {
-  width: 100%;
-}
-
-.table tr:hover {
-  z-index: 1;
-  background-color: #ebefff;
-  
-}
-
-.table td {
-  padding: 12px 0px;
-  border: 2px solid white;
-}
-
-
-.table td:first-child,
-.table th:first-child {
-  position: sticky;
-  left: 0px;
-  z-index: 2;
-  box-shadow: #c32e2e;
-}
-
-.table tr td:nth-child(2),
-.table tr th:nth-child(2) {
-  position: sticky;
-  left: 20px;
-  z-index: 2;
-}
-
-.bottom {
-  display: flex;
-  justify-content: space-between;
-  padding-bottom: 20px;
-}
-
-.pages {
-  display: flex;
-  gap: 10px;
-}
-
-.arrow {
-  color: #777;
-  cursor: pointer;
-}
-
-.numbers {
-  display: flex;
-  gap: 10px;
-}
-
-.active {
-  padding: 2px 9px;
-  border-radius: 5px;
-  border: 1px solid #3f81e0;
-  background-color: #3f81e0;
-  color: white;
-}
-
-.inactive {
-  padding: 2px 9px;
-  border-radius: 5px;
-  border: 1px solid #3f81e0;
-  color: #777;
-}
-
-.lucidePencil {
-  margin-left: 5px;
-  cursor: pointer;
-}
-
-.showingPTag {
-  font-size: 12px;
-  margin: 12px 0px;
-  color: #888;
-}
-
-.cenDiv {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 0px 10px;
-}
-
-.btnnb {
-  background-color: transparent;
-  border-width: 0px;
-  margin: 5px 0px 0px 0px;
-  cursor: pointer;
-}
-
-.lucideCancel {
-  margin-left: 5px;
-  color: #ff0000;
-  cursor: pointer;
-}
-
-.tdStateBlock {
-  text-align: center;
-}
-
-.table tr {
-  transition: background-color 150ms ease; 
-}
-
-@media only screen and (max-width: 1080px) {
-  .top {
-    padding: 15px 0px;
-  }
-
-  .filters {
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    gap: 20px 0;
-    padding-top: 35px;
-  }
-
-  .select {
-    width: 100%;
-  }
-
-  .selectInner {
-    width: 100%;
-    margin-bottom: 15px;
-  }
-
-  .search {
-    width: 100%;
-    height: 50%;
-  }
-
-  .search input {
-    width: 100%;
-    height: 100%;
-  }
-
-  .table th {
-    display: none;
-  }
-
-  .thTag {
-    display: inline;
-  }
-
-  .tableOSTTag {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-evenly;
-    align-items: center;
-    gap: 30px;
-  }
-
-  .table tr:nth-child(1) {
-    display: none;
-  }
-
-  .table tr {
-    display: flex;
-    flex-direction: column;
-    min-width: 250px;
-    border-radius: 5px;
-    border: solid 1px #d3d3d3;
-    overflow: hidden;
-  }
-
-  .table td {
-    border-bottom: 0px solid #ddd;
-    color: black;
-  }
-
-  .table td:first-child {
-    padding-top: 1rem;
-  }
-
-  .table td:last-child {
-    padding-bottom: 1rem;
-  }
-
-  .table td {
-    display: block;
-    padding: 0.5rem 1rem;
-  }
-
-  .table td::before {
-    content: attr(data-cell) ": ";
-    font-weight: 700;
-    text-transform: capitalize;
-    color: black;
-  }
-
-  .filters {
-    display: flex;
-    margin-left: 0px;
-    margin-top: 30px;
-  }
-}
-
-@media screen and (max-width: 450px) {
-  .mDiv {
-    margin-right: 25px;
-  }
-
-  .top {
-    width: 90%;
-  }
-
-  .middle {
-    padding: 0px;
-  }
-
-  .table {
-    padding: 10px 0px;
-    overflow-x: scroll;
-    width: 100%;
-  }
-}
