+++ src/components/MainAdmin/FrontPage.jsx
@@ -64,13 +64,9 @@ export default function FrontPage() {
         <ListBox
           title="Support & Help Desk"
           label="Support and Help Desk Data"
-          redirect="/me/admin/super/support"
+          redirect="/me"
         />
-        <ListBox
-          title="Inventory Statistics"
-          label="Sales data"
-          redirect="/me/admin/super/sales"
-        />
+        <ListBox title="Inventory Statistics" label="Sales data" />
         <ListBox title="Inventory" label="Our Inventory" />
       </div>
     </>
