+++ src/components/StoreVerify/FileInput.jsx
@@ -1,13 +1,10 @@
 import React from "react";
-import { useRef } from "react";
-
-//proptypes
 import PropTypes from "prop-types";
 
-//css
+// css
 import SvCss from "../../Pages/Css/StoreVerify.module.css";
 
-const FileInput = (props) => {
+export default function FileInput(props) {
   const handleClicks = (val) => {
     if (val === "cheque") {
       props.fileInp.current.click();
@@ -19,11 +16,11 @@ const FileInput = (props) => {
   };
 
   return (
-    <div className={SvCss.inputLdivFile}>
-      <p className={SvCss.inputLabel}>{props.label}</p>
-      <div className={SvCss.inputDivFile}>
+    <div className={SvCss.input_ldiv_file}>
+      <p className={SvCss.input_label}>{props.label}</p>
+      <div className={SvCss.input_div_file}>
         <input
-          className={SvCss.inputFile}
+          className={SvCss.input_file}
           type="file"
           name="file"
           placeholder={props.placeholder}
@@ -52,56 +49,12 @@ const FileInput = (props) => {
       </div>
     </div>
   );
-};
-const ImgInputList = (props) => {
-  const fileInp_id = useRef(null);
-  const fileInp_cheque = useRef(null);
-  const fileInp_address = useRef(null);
-  const handleImageCheque = (e) => {
-    props.setImages({ ...props.images, imageUploadCheque: e.target.files[0] });
-  };
-  const handleImageAddress = (e) => {
-    props.setImages({ ...props.images, imageUploadAddress: e.target.files[0] });
-  };
-  const handleImageID = (e) => {
-    props.setImages({ ...props.images, imageUploadID: e.target.files[0] });
-  };
-  return (
-    <>
-      <FileInput
-        label="Upload Cancelled Cheque"
-        placeholder="Cheque "
-        handleImage={handleImageCheque}
-        fileInp={fileInp_cheque}
-        image={props.images.imageUploadCheque}
-        handleClicksValue="cheque"
-      />
-      <FileInput
-        label="Address Proof (GSTIN)"
-        placeholder="Address "
-        handleImage={handleImageAddress}
-        fileInp={fileInp_address}
-        image={props.images.imageUploadAddress}
-        handleClicksValue="address"
-      />
-      <FileInput
-        label="ID Proof (PAN CARD)"
-        placeholder="PAN Card "
-        handleImage={handleImageID}
-        fileInp={fileInp_id}
-        image={props.images.imageUploadID}
-        handleClicksValue="id"
-      />
-    </>
-  );
-};
+}
+
 FileInput.propTypes = {
-  props: PropTypes.shape({
-    label: PropTypes.string,
-    placeholder: PropTypes.string,
-    type1: PropTypes.string,
-    fileInp: PropTypes.object,
-    image: PropTypes.object,
-  }),
+  label: PropTypes.string,
+  placeholder: PropTypes.string,
+  type1: PropTypes.string,
+  fileInp: PropTypes.object,
+  image: PropTypes.object,
 };
-export default ImgInputList;
