+++ src/components/StoreVerify/Heading.jsx
@@ -7,9 +7,9 @@ export default function Heading() {
   return (
     <div className={SvCss.heading}>
       <p className={SvCss.createYourStore}>KYC DATA</p>
-      <div className={SvCss.saveButtons}>
-        <button className={SvCss.saveButtonPurple}>Save</button>
-        <button className={SvCss.saveButtonGreen}>Save & Next</button>
+      <div className={SvCss.save_buttons}>
+        <button className={SvCss.save_button_1}>Save</button>
+        <button className={SvCss.save_button_2}>Save & Next</button>
       </div>
     </div>
   );
