+++ src/components/StoreVerify/PincodeField.jsx
@@ -1,18 +1,15 @@
 import React from "react";
-
-//Proptypes
 import PropTypes from "prop-types";
 
-//axios
+// axios
 import axios from "axios";
 
-//css
+// css
 import SvCss from "../../Pages/Css/StoreVerify.module.css";
 
 const pincodeVerify = async ({
   setData,
   showData,
-  disable,
   setDisable,
   setVerify,
   setError,
@@ -24,14 +21,14 @@ const pincodeVerify = async ({
         State: response.data[0].PostOffice[0].State,
         City: response.data[0].PostOffice[0].Name,
       });
-      setDisable({ ...disable, Pincode: true });
+      setDisable(true);
       setVerify(true);
     };
     const invalidPin = () => {
       setError({
         mainColor: "#FFC0CB",
         secondaryColor: "#FF69B4",
-        symbol: "error",
+        symbol: "pets",
         title: "Check it out",
         text: "Invalid pincode",
         val: true,
@@ -60,10 +57,10 @@ export default function PincodeField({
     <div>
       {showData.Pincode.length >= 6 ? (
         <div className={SvCss.inpDiv}>
-          <p className={SvCss.inputLabel}>Pincode</p>
-          <div className={SvCss.inputDivPincode}>
+          <p className={SvCss.input_label}>Pincode</p>
+          <div className={SvCss.input_div_pincode}>
             <input
-              disabled={disable.Pincode}
+              disabled={disable}
               type="number"
               name="Pincode"
               value={showData.Pincode}
@@ -73,14 +70,13 @@ export default function PincodeField({
                 setData({ ...showData, Pincode: e.target.value });
               }}
             />
-            <div>
+            <div className={SvCss.verify_button_div}>
               <button
-                className={SvCss.verifyButton}
+                className={SvCss.verify_button}
                 onClick={() => {
                   pincodeVerify({
                     setData,
                     showData,
-                    disable,
                     setDisable,
                     setVerify,
                     setError,
@@ -112,10 +108,10 @@ export default function PincodeField({
         </div>
       ) : (
         <div className={SvCss.inpDiv}>
-          <p className={SvCss.inputLabel}>Pincode</p>
-          <div className={SvCss.inputDivPincode}>
+          <p className={SvCss.input_label}>Pincode</p>
+          <div className={SvCss.input_div_pincode}>
             <input
-              disabled={disable.Pincode}
+              disabled={disable}
               type="number"
               name="Pincode"
               value={showData.Pincode}
@@ -133,9 +129,7 @@ export default function PincodeField({
 }
 
 PincodeField.propTypes = {
-  // props: PropTypes.shape({
   showData: PropTypes.object,
   verifyPin: PropTypes.bool,
-  disable: PropTypes.object,
-  // }),
+  disable: PropTypes.bool,
 };
