+++ src/components/StoreVerify/TextInput.jsx
@@ -1,83 +1,13 @@
 import React from "react";
-import { useContext } from "react";
 import PropTypes from "prop-types";
-
-import AuthContext from "../../store/auth-context";
 
 // css
 import SvCss from "../../Pages/Css/StoreVerify.module.css";
 
-export const FssaiField = (props) => {
-  const authCtx = useContext(AuthContext);
-  return (
-    <>
-      {" "}
-      {authCtx.user.category.find(
-        (val) => val.name === "Food & Beverage" || val.name === "Grocery"
-      ) ? (
-        <TextInput
-          type="number"
-          Label="FSSAI Licence NO"
-          showData={props.showData}
-          setData={props.setData}
-          field="FssaiLicence"
-          placeholder="14-digit FSSAI Licence Number"
-        />
-      ) : (
-        ""
-      )}
-    </>
-  );
-};
-export const GrpTextInput = (props) => {
-  return (
-    <>
-      <TextInput
-        type="text"
-        Label="First Name"
-        showData={props.showData}
-        setData={props.setData}
-        field="FirstName"
-        placeholder="John"
-      />{" "}
-      <TextInput
-        type="text"
-        Label="Last Name"
-        showData={props.showData}
-        setData={props.setData}
-        field="LastName"
-        placeholder="David"
-      />{" "}
-      <TextInput
-        type="text"
-        Label="Legal Name"
-        showData={props.showData}
-        setData={props.setData}
-        field="LegalName"
-        placeholder="John David"
-      />
-      <TextInput
-        type="email"
-        Label="Email ID"
-        showData={props.showData}
-        setData={props.setData}
-        field="EmailID"
-        placeholder="Enter your email"
-      />{" "}
-      <TextInput
-        type="date"
-        Label="DOB"
-        showData={props.showData}
-        setData={props.setData}
-        field="DOB"
-      />
-    </>
-  );
-};
-const TextInput = (props) => {
+export default function TextInput(props) {
   return (
     <div className={SvCss.inpDiv}>
-      <p className={SvCss.inputLabel}>{props.Label}</p>
+      <p className={SvCss.input_label}>{props.Label}</p>
       <input
         type={props.type}
         name="days"
@@ -90,15 +20,12 @@ const TextInput = (props) => {
       />
     </div>
   );
-};
+}
 
 TextInput.propTypes = {
-  props: PropTypes.shape({
-    Label: PropTypes.string,
-    type: PropTypes.string,
-    field: PropTypes.string,
-    placeholder: PropTypes.string,
-    showData: PropTypes.object,
-  }),
+  Label: PropTypes.string,
+  type: PropTypes.string,
+  field: PropTypes.string,
+  placeholder: PropTypes.string,
+  showData: PropTypes.object,
 };
-export default TextInput;
