+++ src/components/StoreVerify/TimingField.jsx
@@ -1,13 +1,14 @@
 import React from "react";
 import PropTypes from "prop-types";
-//css
+
+// css
 import SvCss from "../../Pages/Css/StoreVerify.module.css";
 
 export default function TimingField(props) {
   return (
-    <div className={SvCss.timingLargeDiv}>
-      <p className={SvCss.inputLabel}>Store Timing</p>
-      <div className={SvCss.timingSmallDiv}>
+    <div className={SvCss.timing_large_div}>
+      <p className={SvCss.input_label}>Store Timing</p>
+      <div className={SvCss.timing_small_div}>
         <input
           type="time"
           name="Store_Timing"
@@ -40,7 +41,5 @@ export default function TimingField(props) {
 }
 
 TimingField.propTypes = {
-  props: PropTypes.shape({
-    showData: PropTypes.object,
-  }),
+  showData: PropTypes.object,
 };
