--- src/components/StoreVerify/getLocation.jsx
@@ -1,19 +1,0 @@
-import React from "react";
-
-export default function geolocation(showData, setData) {
-  const successCallback = (position) => {
-    setData({
-      ...showData,
-      gps: {
-        latitude: position.coords.latitude,
-        longitude: position.coords.longitude,
-      },
-    });
-  };
-
-  const errorCallback = (error) => {
-    console.log(error);
-  };
-
-  navigator.geolocation.getCurrentPosition(successCallback, errorCallback);
-}
